requirements.txt
Flask==3.0.3
gunicorn==21.2.0
PyMuPDF==1.24.9
python-dotenv==1.0.1
twilio==9.0.5
requests==2.32.4
Pillow==10.4.0
pyzbar==0.1.9
